Blocco di modifiche importante ....con aggiunta prototipo gestione commessa

+++ core/promogest/ui/AnagraficaComplessa.py
@@ -368,7 +368,8 @@ class Anagrafica(GladeWidget):
             return
 
         dao = self.filter.getSelectedDao()
-        dao.delete()
+        if dao:
+            dao.delete()
         self.filter.refresh()
         self.htmlHandler.setDao(None)
         self.setFocus()
